requirements.txt
streamlit
google-generativeai
protobuf<4.0.0

